chore: 채팅창 메시지 영역 컴포넌트로 분리
